[CHORE] AppleSignInDelegate file 분리, Label Height extension
