+++ Puhaha/Puhaha/Controllers/SignInViewController.swift
@@ -6,13 +6,11 @@
 //
 
 import UIKit
-import FirebaseAuth
 import AuthenticationServices
-import CryptoKit
 
 class SignInViewController: UIViewController {
     
-    private var currentNonce: String?
+    public var currentNonce: String?
     
     private let titleLabel: UILabel = {
       let label = UILabel()
@@ -73,121 +71,5 @@ class SignInViewController: UIViewController {
 extension SignInViewController {
     @objc func appleLoginButtonTapped() {
         startSignInWithAppleFlow()
-    }
-}
-
-extension SignInViewController: ASAuthorizationControllerDelegate {
-    func authorizationController(controller: ASAuthorizationController, didCompleteWithAuthorization authorization: ASAuthorization) {
-        
-        if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
-            guard let nonce = currentNonce else {
-                fatalError("Invalid state: A login callback was received, but no login request was sent.")
-            }
-            guard let appleIDToken = appleIDCredential.identityToken else {
-                print("Unable to fetch identity token")
-                return
-            }
-            guard let idTokenString = String(data: appleIDToken, encoding: .utf8) else {
-                print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
-                return
-            }
-            
-            let credential = OAuthProvider.credential(withProviderID: "apple.com",
-                                                      idToken: idTokenString,
-                                                      rawNonce: nonce)
-            
-            Auth.auth().signIn(with: credential) { _, error in
-                if let error = error {
-                    print("Error Apple sign in: %@", error)
-                    return
-                }
-            }
-            
-            let mainTabViewController = MainTabViewController()
-            self.navigationController?.pushViewController(mainTabViewController, animated: true)
-        }
-    }
-}
-
-extension SignInViewController {
-    func startSignInWithAppleFlow() {
-        let nonce = randomNonceString()
-        currentNonce = nonce
-        let appleIDProvider = ASAuthorizationAppleIDProvider()
-        let request = appleIDProvider.createRequest()
-        request.requestedScopes = [.fullName, .email]
-        request.nonce = sha256(nonce)
-        
-        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
-        authorizationController.delegate = self
-        authorizationController.presentationContextProvider = self
-        authorizationController.performRequests()
-    }
-    
-    private func sha256(_ input: String) -> String {
-        let inputData = Data(input.utf8)
-        let hashedData = SHA256.hash(data: inputData)
-        let hashString = hashedData.compactMap {
-            return String(format: "%02x", $0)
-        }.joined()
-        return hashString
-    }
-    
-    private func randomNonceString(length: Int = 32) -> String {
-        precondition(length > 0)
-        let charset: [Character] = Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-        var result = ""
-        var remainingLength = length
-        
-        while remainingLength > 0 {
-            let randoms: [UInt8] = (0 ..< 16).map { _ in
-                var random: UInt8 = 0
-                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
-                if errorCode != errSecSuccess {
-                    fatalError("Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)")
-                }
-                return random
-            }
-            
-            randoms.forEach { random in
-                if remainingLength == 0 {
-                    return
-                }
-                if random < charset.count {
-                    result.append(charset[Int(random)])
-                    remainingLength -= 1
-                }
-            }
-        }
-        return result
-    }
-}
-
-extension SignInViewController: ASAuthorizationControllerPresentationContextProviding {
-    func presentationAnchor(for controller: ASAuthorizationController) -> ASPresentationAnchor {
-        return self.view.window!
-    }
-}
-
-extension UILabel {
-    func setLineHeight(lineHeight: CGFloat) {
-        let paragraphStyle = NSMutableParagraphStyle()
-        paragraphStyle.lineSpacing = 1.0
-        paragraphStyle.lineHeightMultiple = lineHeight
-        paragraphStyle.alignment = self.textAlignment
-
-        let mutableAttributedString = NSMutableAttributedString()
-        if (self.attributedText != nil) {
-            mutableAttributedString.append(self.attributedText ?? NSMutableAttributedString() )
-        } else {
-            mutableAttributedString.append(NSMutableAttributedString(string: self.text ?? ""))
-            mutableAttributedString.addAttribute(NSAttributedString.Key.font,
-                                                 value: self.font,
-                                                 range: NSMakeRange(0, mutableAttributedString.length))
-        }
-        mutableAttributedString.addAttribute(NSAttributedString.Key.paragraphStyle,
-                                             value: paragraphStyle,
-                                             range: NSMakeRange(0, mutableAttributedString.length))
-        self.attributedText = mutableAttributedString
     }
 }
